Mon, Jun 17, 2024, 2:56 AM +01:00

+++ finance/app.py
@@ -49,7 +49,7 @@ def index():
                           'worth': int(user_shares[i]["shares"]) * price["price"]
                           })
     for i in range(len(user_data)):
-        total += int(user_data[i]["worth"])
+        total += float(user_data[i]["worth"])
     total += user[0]["cash"]
     return render_template("index.html", user_data=user_data, cash=user[0]["cash"], total=total)
